Adding a base class of oembedresponse, so we can handle string values for dimentions
