+++ src/Umbraco.Core/Media/EmbedProviders/OEmbedResponseWithStringDimensions.cs
@@ -1,0 +1,9 @@
+using System.Runtime.Serialization;
+
+namespace Umbraco.Cms.Core.Media.EmbedProviders;
+
+/// <summary>
+///     Wrapper class for OEmbed response with width and height as string values.
+/// </summary>
+[DataContract]
+public class OEmbedResponseWithStringDimensions : OEmbedResponseBase<string>;
